+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0-alpha] - 2021-10-13
 ### Added
 - Eslint
 - Dependabot
@@ -57,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Inital commit
 
 [Unreleased]: https://github.com/kayano-bot/kayano/compare/stable...development
+[1.1.0]: https://github.com/kayano-bot/kayano/compare/v1.0.0-alpha...v1.1.0-alpha
 [1.0.0]: https://github.com/kayano-bot/kayano/compare/v0.0.2...v1.0.0-alpha
 [0.0.2]: https://github.com/kayano-bot/kayano/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/kayano-bot/kayano/commits/v0.0.1